BUG/CLN: make xs a generic method
