+++ doc/source/release.rst
@@ -129,6 +129,7 @@ Bug Fixes
   - Bug in plotting.scatter_matrix function. Wrong alignment among diagonal
     and off-diagonal plots, see (:issue:`5497`).
   - Regression in Series with a multi-index via ix (:issue:`6018`)
+  - Bug in Series.xs with a multi-index (:issue:`6018`)
 
 pandas 0.13.0
 -------------
