+++ pandas/core/series.py
@@ -445,11 +445,6 @@ class Series(generic.NDFrame):
 
         return values
 
-    def _xs(self, key, axis=0, level=None, copy=True):
-        return self.__getitem__(key)
-
-    xs = _xs
-
     def _ixs(self, i, axis=0):
         """
         Return the i-th value or values in the Series by location
